+++ software/ors-amarisoft/changelog.rst
@@ -11,3 +11,11 @@ Version 1.0.312 (2023-03-20)
 
 * Add promise to test if reception is saturated
 * Add IMSI in connection parameters when SIM gets attached
+
+Version 1.0.316 (2023-04-14)
+-------------
+
+* Remove enb-epc, gnb-epc and epc software types, the software types are now:
+    - enb
+    - gnb
+    - core-network (replaces epc software type)